moving all client scripts into /src
