--- public/scripts/defer.js
@@ -1,3 +1,0 @@
-if (window.matchMedia('(hover: hover)').matches) {
-    import('/scripts/konami.js').then(mod => mod.init());
-}
