+++ src/components/BaseLayout.astro
@@ -27,4 +27,4 @@ const { wrapper: Wrapper = Fragment, container = false } = Astro.props;
         document.body.classList.add('js');
     })
 </script>
-<script type="module" src="/scripts/defer.js"></script>
+<script type="module" src="/src/scripts/polyfill.ts"></script>
